tests: Add GPU-AV + RT buffer val + desc heap test
