+++ tests/unit/gpu_av_ray_tracing.cpp
@@ -4400,178 +4400,3 @@ TEST_F(NegativeGpuAVRayTracing, OpReportIntersectionKHRHitKindOutOfRange) {
     ASSERT_GT(debug_buffer_ptr[0], 0u) << "Intersection shader was never invoked";
     debug_buffer.Memory().Unmap();
 }
-
-// https://github.com/KhronosGroup/Vulkan-ValidationLayers/issues/12657
-TEST_F(NegativeGpuAVRayTracing, DISABLED_InvalidBlasReference1DescriptorHeap) {
-    TEST_DESCRIPTION(
-        "Validate an invalid BLAS reference in a TLAS build - first element BLAS ref invalid, subsequent ones valid."
-        "Trace a ray into the built TLAS to confirm it was built correctly without the invalid ref but with the valid ones."
-        "Use descriptor heaps");
-
-    RETURN_IF_SKIP(CheckSlangSupport());
-
-    SetTargetApiVersion(VK_API_VERSION_1_2);
-    vkt::DescriptorHeap::AddDescriptorHeapRequirements(*this);
-    AddRequiredExtensions(VK_KHR_MAINTENANCE_4_EXTENSION_NAME);
-    AddRequiredFeature(vkt::Feature::rayTracingPipeline);
-    AddRequiredFeature(vkt::Feature::accelerationStructure);
-    AddRequiredFeature(vkt::Feature::bufferDeviceAddress);
-    AddRequiredFeature(vkt::Feature::maintenance4);
-    AddRequiredFeature(vkt::Feature::shaderInt64);
-
-    VkValidationFeaturesEXT validation_features = GetGpuAvValidationFeatures();
-    RETURN_IF_SKIP(InitFrameworkForRayTracingTest(&validation_features));
-    if (!CanEnableGpuAV(*this)) {
-        GTEST_SKIP() << "Requirements for GPU-AV are not met";
-    }
-    RETURN_IF_SKIP(InitState());
-    InitRenderTarget();
-
-    vkt::as::GeometryKHR cube(vkt::as::blueprint::GeometryCubeOnDeviceInfo(*m_device));
-    vkt::as::BuildGeometryInfoKHR cube_blas = vkt::as::blueprint::BuildGeometryInfoOnDeviceBottomLevel(*m_device, std::move(cube));
-
-    m_command_buffer.Begin();
-    cube_blas.BuildCmdBuffer(m_command_buffer);
-    m_command_buffer.End();
-
-    m_default_queue->Submit(m_command_buffer);
-    m_device->Wait();
-
-    std::vector<vkt::as::GeometryKHR> cube_instances(1);
-    cube_instances[0].SetType(vkt::as::GeometryKHR::Type::Instance);
-
-    VkAccelerationStructureInstanceKHR cube_instance_1{};
-    cube_instance_1.transform.matrix[0][0] = 1.0f;
-    cube_instance_1.transform.matrix[1][1] = 1.0f;
-    cube_instance_1.transform.matrix[2][2] = 1.0f;
-    cube_instance_1.transform.matrix[0][3] = 50.0f;
-    cube_instance_1.transform.matrix[1][3] = 0.0f;
-    cube_instance_1.transform.matrix[2][3] = 0.0f;
-    cube_instance_1.mask = 0xff;
-    cube_instance_1.instanceCustomIndex = 0;
-    // Cube instance 1 will be associated to closest hit shader 1
-    cube_instance_1.instanceShaderBindingTableRecordOffset = 0;
-    cube_instances[0].AddInstanceDeviceAccelStructRef(*m_device, cube_blas.GetDstAS()->handle(), cube_instance_1);
-    cube_instances[0].AddInstanceDeviceAccelStructRef(*m_device, cube_blas.GetDstAS()->handle(), cube_instance_1);
-
-    VkAccelerationStructureInstanceKHR cube_instance_2{};
-    cube_instance_2.transform.matrix[0][0] = 1.0f;
-    cube_instance_2.transform.matrix[1][1] = 1.0f;
-    cube_instance_2.transform.matrix[2][2] = 1.0f;
-    cube_instance_2.transform.matrix[0][3] = 0.0f;
-    cube_instance_2.transform.matrix[1][3] = 0.0f;
-    cube_instance_2.transform.matrix[2][3] = 50.0f;
-    cube_instance_2.mask = 0xff;
-    cube_instance_2.instanceCustomIndex = 0;
-    // Cube instance 2 will be associated to closest hit shader 1
-    cube_instance_2.instanceShaderBindingTableRecordOffset = 0;
-
-    cube_instances[0].AddInstanceDeviceAccelStructRef(*m_device, cube_blas.GetDstAS()->handle(), cube_instance_2);
-
-    std::vector<vkt::as::BuildGeometryInfoKHR> tlas_build_info;
-    {
-        vkt::as::BuildGeometryInfoKHR tlas = vkt::as::blueprint::CreateTLAS(*m_device, std::move(cube_instances));
-        tlas_build_info.emplace_back(std::move(tlas));
-        m_command_buffer.Begin();
-        vkt::as::BuildAccelerationStructuresKHR(m_command_buffer, tlas_build_info);
-        m_command_buffer.End();
-
-        m_default_queue->Submit(m_command_buffer);
-        m_device->Wait();
-    }
-    // Buffer used to count invocations for the 3 shaders
-    vkt::Buffer debug_buffer(*m_device, 3 * sizeof(uint32_t), VK_BUFFER_USAGE_STORAGE_BUFFER_BIT | VK_BUFFER_USAGE_TRANSFER_DST_BIT,
-                             kHostVisibleMemProps);
-    auto debug_buffer_ptr = static_cast<uint32_t*>(debug_buffer.Memory().Map());
-    std::memset(debug_buffer_ptr, 0, (size_t)debug_buffer.CreateInfo().size);
-
-    vkt::DescriptorHeap desc_heap(*this);
-    desc_heap.CreateResourceHeap(1024, true);
-    const VkDeviceSize as_heap_offset = desc_heap.WriteAccelerationStructureDescriptor(*tlas_build_info[0].GetDstAS());
-
-    const VkDeviceSize as_descriptor_size =
-        vk::GetPhysicalDeviceDescriptorSizeEXT(m_device->Physical(), VK_DESCRIPTOR_TYPE_ACCELERATION_STRUCTURE_KHR);
-
-    VkDescriptorSetAndBindingMappingEXT mapping = vku::InitStructHelper();
-    mapping = MakeSetAndBindingMapping(0, 0);
-    mapping.source = VK_DESCRIPTOR_MAPPING_SOURCE_HEAP_WITH_SHADER_RECORD_INDEX_EXT;
-    mapping.sourceData.shaderRecordIndex = {};
-    mapping.sourceData.shaderRecordIndex.heapOffset = (uint32_t)as_heap_offset;
-    mapping.sourceData.shaderRecordIndex.shaderRecordOffset = 0;
-    mapping.sourceData.shaderRecordIndex.heapIndexStride = (uint32_t)as_descriptor_size;
-    mapping.sourceData.shaderRecordIndex.heapArrayStride = (uint32_t)as_descriptor_size;
-    VkShaderDescriptorSetAndBindingMappingInfoEXT mapping_info = vku::InitStructHelper();
-    mapping_info.mappingCount = 1u;
-    mapping_info.pMappings = &mapping;
-
-    const char* slang_shader = R"slang(
-        [[vk::binding(0, 0)]] uniform RaytracingAccelerationStructure tlas;
-        [[vk::binding(1, 0)]] RWStructuredBuffer<uint32_t> debug_buffer;
-
-        struct RayPayload {
-            uint4 payload;
-            float3 hit;
-        };
-
-        [shader("raygeneration")]
-        void rayGenShader()
-        {
-            InterlockedAdd(debug_buffer[0], 1);
-            RayPayload ray_payload = {};
-            RayDesc ray;
-            ray.TMin = 0.01;
-            ray.TMax = 1000.0;
-
-            // Will hit cube 1
-            ray.Origin = float3(0,0,0);
-            ray.Direction = float3(1,0,0);
-            TraceRay(tlas, RAY_FLAG_NONE, 0xff, 0, 0, 0, ray, ray_payload);
-        }
-
-        [shader("miss")]
-        void missShader(inout RayPayload payload)
-        {
-            InterlockedAdd(debug_buffer[1], 1);
-            payload.hit = float3(0.1, 0.2, 0.3);
-        }
-
-        [shader("closesthit")]
-        void closestHitShader(inout RayPayload payload, in BuiltInTriangleIntersectionAttributes attr)
-        {
-            InterlockedAdd(debug_buffer[2], 1);
-            const float3 barycentric_coords = float3(1.0f - attr.barycentrics.x - attr.barycentrics.y, attr.barycentrics.x,
-                attr.barycentrics.y);
-            payload.hit = barycentric_coords;
-        }
-    )slang";
-
-    vkt::rt::Pipeline pipeline(*this, m_device);
-    pipeline.AddCreateInfoFlags2(VK_PIPELINE_CREATE_2_DESCRIPTOR_HEAP_BIT_EXT);
-    pipeline.SetShaderRecordSize(3 * sizeof(uint32_t));
-    pipeline.AddSlangRayGenShader(slang_shader, "rayGenShader", nullptr, &mapping_info);
-    pipeline.AddSlangMissShader(slang_shader, "missShader", nullptr, &mapping_info);
-    pipeline.AddSlangClosestHitShader(slang_shader, "closestHitShader", nullptr, &mapping_info);
-
-    pipeline.Build();
-
-    uint32_t vec[3] = {0, 2, 4};
-    pipeline.UpdateRayGenShaderRecord(0, vec, 3 * sizeof(uint32_t));
-
-    m_command_buffer.Begin();
-
-    desc_heap.BindResourceHeap(m_command_buffer);
-    vk::CmdBindPipeline(m_command_buffer, VK_PIPELINE_BIND_POINT_RAY_TRACING_KHR, pipeline);
-
-    vkt::rt::TraceRaysSbt sbt_ray_gen_1 = pipeline.GetTraceRaysSbt(0);
-    vk::CmdTraceRaysKHR(m_command_buffer, &sbt_ray_gen_1.ray_gen_sbt, &sbt_ray_gen_1.miss_sbt, &sbt_ray_gen_1.hit_sbt,
-                        &sbt_ray_gen_1.callable_sbt, 1, 1, 1);
-
-    m_command_buffer.End();
-
-    m_default_queue->SubmitAndWait(m_command_buffer);
-
-    // Make sure expected ray tracing setup worked, indicating the TLAS was correctly built
-    ASSERT_EQ(debug_buffer_ptr[0], 1);
-    ASSERT_EQ(debug_buffer_ptr[1], 0);
-    ASSERT_EQ(debug_buffer_ptr[2], 1);
-}
